requirements.txt
openai==1.x.x
python-dotenv==1.x.x
pandas==2.x.x
numpy==1.x.x
seaborn==0.x.x
matplotlib==3.x.x